Remove out of date "implemented" column
The dates were wrong on the column of implemented modules. Also, it's not terrible helpful, so I removed it.
These modules were also all described as being migrated out, which isn't correct for all of them. Broke into two tables, one for lean core migrations, the other for community modules.
Long term those should be tracked with a [directory](https://github.com/microsoft/react-native-windows/issues/3718), but for now while the list is short this is still helpful.