Update NativeModuleSample to use 0.64 library templates and add upgrade instructions (#536)
NativeModuleSample was created before the cli templates were added to RNW 0.64 and exposed through `npx react-native-windows-init --projectType lib`.

This PR updates the samples to align with the template, to facilitate upgrades to newer versions of RNW using the newer templates. It also adds setup, build, and upgrade steps to the readmes.